Mitigations to reduce downstream library sizes.
This commit adds some explicit template instantiations to decrease
library sizes for downstream packages.  In particular, we instantiate
much of the ParameterSet-get functionality for:

 - bool
 - int
 - unsigned
 - float
 - double
 - std::string
 - fhicl::ParameterSet

We also add instantiations for std::vectors of the above types.

Lastly, we have moved the implementations of some destructors and
other functions to implementation files to reduce downstream
code-bloat.